requirements.txt
azure-cognitiveservices-vision-computervision
flask